all the maps are working now
however theres some ID bullshit that i had to do when printing the map

that is going to come back to bite me later

really, i should adjust the ids. mcgrew says that the highest bit in the nibble is ignored for dungeons, and that is really what is causing my problems. if i fix that one, i shouldn't have to do all the weird id stuff when printing.